trying lag compensation

+++ Shared/Items/Weapons/Weapon.cs
@@ -9,7 +9,14 @@ namespace Fanior.Shared
 {
     public abstract class Weapon
     {
-        public Weapon(bool autoFire, int reloadTime, int shotSpeed, double damage)
+        /// <summary>
+        /// Weapon of a character
+        /// </summary>
+        /// <param name="autoFire">Whether player must press key each time or it's automatic</param>
+        /// <param name="reloadTime">number of frames it takes to reload</param>
+        /// <param name="shotSpeed">What it says...</param>
+        /// <param name="damage">Damage of the created shot</param>
+        public Weapon(bool autoFire, double reloadTime, int shotSpeed, double damage)
         {
             this.reloadTime = reloadTime;
             this.shotSpeed = shotSpeed;
@@ -17,7 +24,8 @@ namespace Fanior.Shared
             this.autoFire = autoFire;
         }
         public bool reloaded = true;
-        public int reloadTime;
+        //number of frames it takes to reload
+        public double reloadTime;
         [JsonProperty]
         protected int shotSpeed;
         [JsonProperty]
